prepping details section to not use routes
